// src/pages/user/institution/types.ts
export type ServiceScheduleItem = {
  day: string;
  time: string;
  note: string;
};

export type LeaderItem = {
  name: string;
  role: string;
  specialty: string;
};

export type InstitutionSeriesItem = {
  title: string;
  meta: string;
};

export type InstitutionEventItem = {
  title: string;
  date: string;
  type: string;
  price: string;
};